requirements.txt
selenium==4.11.2
python-dotenv